extract location + death handling to CharacterPhysicalModel
